split the testcases into seperate tests
